% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Beverly}
\alias{"Dr.}
\alias{Beverly}
\alias{Crusher"}
\title{Talk like Commander Beverly Cheryl Crusher, MD, (née Howard)}
\usage{
Beverly()
}
\value{
a phrase
}
\description{
Talk with the sensible and kind insight that
a Chief Medical Officer on a starship develops.
}
\seealso{
Other Darmok: \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Guinan}},
  \code{\link{Picard}}, \code{\link{Pulaski}},
  \code{\link{Q}}, \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

