% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Computer}
\alias{Computer}
\title{Computer voice}
\usage{
Computer()
}
\value{
a phrase
}
\description{
Always ready to tell us about schip problems
or anything really. The Siri of the 24th century,
but you know, working. Almost never misanderstanding you.
}
\examples{
Guinan()
}
\seealso{
Other Darmok: \code{\link{Beverly}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

