% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Data}
\alias{Data}
\title{Talk like Commander Data}
\usage{
Data()
}
\value{
a phrase
}
\description{
Like an Amazon Echo in clothes, but a fully functional
one if we have to believe Yara. This function comes
with inspirational sentances like: " Deck seventeen.",
or " Captain. There are just in the Federation,
eighty-one different variations of signing."
}
\examples{
Data()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

