% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Dathon}
\alias{Darmok}
\alias{Dathon}
\alias{Jalad}
\title{Dathon}
\usage{
Dathon()

Darmok()
}
\value{
a phrase
}
\description{
The enigmatic Tamarians also known as
the Children of Tama are called "incomprehensible.
Although meetings with this species has not lead
to conflicts communications was never possible.
Dathon, the chief of this folk, speaks in riddles.
"Shaka. When the walls fell", "Temba. His arms wide."
}
\examples{
Dathon()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Guinan}},
  \code{\link{Picard}}, \code{\link{Pulaski}},
  \code{\link{Q}}, \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

