% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Geordi}
\alias{Geordi}
\title{Speak like Geordi}
\usage{
Geordi()
}
\value{
a phrase
}
\description{
Truly man with a vision. Geordi can
see in many frequencies. Expect a lot
of technobabble and a few creepy romance words.
}
\examples{
Geordi()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Guinan}},
  \code{\link{Picard}}, \code{\link{Pulaski}},
  \code{\link{Q}}, \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

