% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Guinan}
\alias{Guinan}
\title{Speak Like Guinan}
\usage{
Guinan()
}
\value{
a phrase
}
\description{
A barkeeper with an enourmous gun
and lots of huge hats. A being outside
of time and incredibly wise.
}
\examples{
Guinan()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

