% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Picard}
\alias{JPL}
\alias{Jeanluc}
\alias{Picard}
\title{Return a phrase that Picard uttered}
\usage{
Picard()
}
\value{
a phrase
}
\description{
Be as eloquent as our favorite archaeologist, diplomat and
decorated Starfleet officer Jean-Luc Picard, or as we
tend to say "JLP". The startrek package does not help
you in becoming a Starfleet officer, or archaeologist
for that matter. But perhaps you will use the same
phrasing when you order your Tea Earl Grey.
As Picard once said : "Darmok and Jalad... at Tanagra".
}
\examples{
Picard()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Guinan}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

