% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Pulaski}
\alias{Pulaski}
\title{Talk like Pulaski}
\usage{
Pulaski()
}
\value{
a phrase
}
\description{
Rough around the edges, probably an alcoholic
but a good Chief Medical Officer in absence of
Beverly. Also slightly racist towards Data
}
\examples{
Pulaski()
}
\seealso{
Beverly Data

Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Q}}, \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

