% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Q}
\alias{Q}
\title{"Q"}
\usage{
Q()
}
\value{
a phrase
}
\description{
Incredibly annoying character.
}
\examples{
Q()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Guinan}},
  \code{\link{Picard}}, \code{\link{Pulaski}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}, \code{\link{Worf}}
}

