% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{random_tng_character}
\alias{random_tng_character}
\title{Call a random phrase from Q, Worf, Wesley, Picard, Data or Guinan}
\usage{
random_tng_character()
}
\description{
Usefull if you want a phrase at the end of a long process.
}
\examples{
random_tng_character()
}

