% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Riker}
\alias{"Number}
\alias{One"}
\alias{Riker}
\alias{Will}
\title{William (Will) "Number One" Riker}
\usage{
Riker()
}
\value{
a phrase
}
\description{
The original leaning man. Riker has a tendency
to solve diplomatic disputes by sleeping with
anyone around. The man, the legend.
}
\examples{
Riker()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Troi}}, \code{\link{Wesley}},
  \code{\link{Worf}}
}

