% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{startrek}
\alias{TNG}
\alias{startrek}
\alias{startrek-package}
\title{Star trek package write like characters spoke}
\description{
This incredibly unuseful package makes you talk like
the main characters from Star Trek The  Next Generation.
call functions with the name of the characters
`Picard()` , `Geordi()`, etc
Characters included are Picard, Beverly, Data,
Geordi, Guinan, Computer, Pulaski, Q, Riker, Troi,
Wesley, Worf and also Dathon.
}

