% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_trek_package.R
\docType{package}
\name{Startrekpackage}
\alias{"Star}
\alias{"TNG}
\alias{Startrekpackage}
\alias{Startrekpackage-package}
\alias{Trek"}
\alias{package"}
\title{Star Trek Package}
\description{
This package will give you: for every main character
in the Star Trek the Next Generation (TNG) a random
selection of phrases. All the phrases have been
lifted from the original Scripts.
}
\details{
All functions are called after the Character. So
just call Pulaski() Or Picard().

This is where you should actually put a link or something
to the original source of the document and the dataset.
}

