% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{TNGvoices}
\alias{TNGvoices}
\title{All the utterances of all the characters in TNG}
\format{a dataframe with every utterance.}
\source{
\url{https://github.com/rmhogervorst/TNG}

\url{http://www.st-minutiae.com/resources/scripts/}
}
\usage{
TNGvoices
}
\description{
Like almost every episode.
}
\keyword{datasets}

