% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Troi}
\alias{Troi}
\title{Troi}
\usage{
Troi()
}
\value{
a phrase
}
\description{
An empath, Bezoid? Anyway, she can say
inspirational stuff such as:
" What the future holds.
No one can know.
But forward we look And Forward you go..."
}
\examples{
Troi()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Wesley}},
  \code{\link{Worf}}
}

