% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Wesley}
\alias{"the}
\alias{"young}
\alias{Wesley}
\alias{boy"}
\alias{the_boy}
\alias{wesley"}
\title{The boy}
\usage{
Wesley()

the_boy()
}
\value{
a phrase
}
\description{
Son of Beverly, eager to work. A warpdrive genius.
Incredibly annoying.
}
\examples{
Wesley()
}
\seealso{
Beverly

Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Geordi}}, \code{\link{Guinan}},
  \code{\link{Picard}}, \code{\link{Pulaski}},
  \code{\link{Q}}, \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Worf}}
}

