% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{Worf}
\alias{Worf}
\title{Leutenant Commander Worf}
\usage{
Worf()
}
\value{
a phrase
}
\description{
The man, the Klingon, has honor
has the metal thingy around him.
Killer chef with a Bathleth.
Worst security officer on the show, I mean
really how many times has the Enterprise been
taken over? Come on!
}
\examples{
Worf()
}
\seealso{
Other Darmok: \code{\link{Beverly}},
  \code{\link{Computer}}, \code{\link{Data}},
  \code{\link{Dathon}}, \code{\link{Geordi}},
  \code{\link{Guinan}}, \code{\link{Picard}},
  \code{\link{Pulaski}}, \code{\link{Q}},
  \code{\link{Riker}}, \code{\link{Troi}},
  \code{\link{Wesley}}
}

