% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{TNGvoices}
\alias{TNGvoices}
\title{All the utterances of TNG}
\format{a dataframe with every utterance.}
\source{
\url{blablba}
}
\usage{
TNGvoices
}
\description{
Like almost every episode.
}
\keyword{datasets}

